Schema Fields:
fullName:{: This defines a field named fullName in the schema.
type:String,: Specifies that the type of data stored in this field should be a string.
require:true: Indicates that this field is required, meaning it must be provided when creating a new document.
email:{: This defines a field named email in the schema.
type:String,: Specifies that the type of data stored in this field should be a string.
require:true: Indicates that this field is required.
password:{: This defines a field named password in the schema.
type:String,: Specifies that the type of data stored in this field should be a string.
require:true: Indicates that this field is required.