Creating the User Model:
export const User = mongoose.model("User", userSchema);: This line creates a model named User based on the userSchema. The mongoose.model function takes two arguments: the name of the model and the schema to use. It returns a Mongoose model that can be used to interact with the users collection in MongoDB. The model is then exported for use in other parts of the application.
In summary, this code defines a Mongoose schema for a User collection, specifying that each user document must have a fullName, email, and password, all of which are required string fields. It also enables automatic timestamping for document creation and updates. Finally, it creates and exports a Mongoose model based on this schema, which can be used to perform CRUD operations on the User collection in MongoDB.