Reducers:
Within the reducers property, each key represents a Redux action type, and its corresponding value is a reducer function that modifies the state based on the action payload:
setUser: Takes state and action as parameters. It updates state.user with the payload received via action.payload. This action is typically dispatched when a user logs in or when user data is fetched from an API.
setLoading: Takes state and action as parameters. It updates state.isLoading with the payload received via action.payload. This action is used to update the loading state, indicating whether data is currently being fetched or not.