Event Handler:
const handleOpen = () => { ... }:
Defines an event handler function handleOpen that dispatches two actions:
getId(movieId): Dispatches an action with the movieId, indicating which movie's details should be loaded.
setOpen(true): Dispatches an action to set the open state to true, indicating that a movie details dialog or modal should be opened.